Let's try moving chat templates out of IDEFICS and into the generic ProcessorMixin
